check autUser

+++ routes/creatorProfiles.js
@@ -5,9 +5,17 @@ const authenticate = passport.authorize('jwt', { session: false })
 
 router.get('/creator/:id', authenticate, (req, res, next) => {
     const id = req.params.id
+    userId = req.account._id
     CreatorProfile.findById(id)
       .then((creatorProfile) => {
         if (!creatorProfile) { return next() }
+
+        if(creatorProfile.user._id !== userId){
+          const error = new Error('Unauthorized')
+          error.status = 401
+          return next(error)
+        }
+
         res.status = 200
         res.json(creatorProfile)
       })
@@ -15,9 +23,16 @@ router.get('/creator/:id', authenticate, (req, res, next) => {
   })
   .post('/creator', authenticate, (req, res, next) => {
     let newCreatorProfile = req.body
-
+    userId = req.account._id
     CreatorProfile.create(newCreatorProfile)
       .then((creatorProfile) => {
+
+        if(creatorProfile.user._id !== userId){
+          const error = new Error('Unauthorized')
+          error.status = 401
+          return next(error)
+        }
+
         res.status = 201
         res.json(creatorProfile)
       })
@@ -26,10 +41,17 @@ router.get('/creator/:id', authenticate, (req, res, next) => {
   .put('/creator/:id', authenticate, (req, res, next) => {
     const creatorProfileId = req.params.id
     let updaCreatorProfile = req.body
-
+    userId = req.account._id
     CreatorProfile.findOneAndUpdate(creatorProfileId, updaCreatorProfile)
     .then((creatorProfile) => {
       if (!creatorProfile) { return next() }
+
+      if(creatorProfile.user._id !== userId){
+        const error = new Error('Unauthorized')
+        error.status = 401
+        return next(error)
+      }
+
       res.status = 200
       res.json(creatorProfile)
     })
@@ -38,24 +60,43 @@ router.get('/creator/:id', authenticate, (req, res, next) => {
   .patch('/creator/:id', authenticate, (req, res, next) => {
     const creatorProfileId = req.params.id
     let updaCreatorProfile = req.body
-
+    userId = req.account._id
     CreatorProfile.findOneAndUpdate(creatorProfileId, updaCreatorProfile)
     .then((creatorProfile) => {
       if (!creatorProfile) { return next() }
+
+      if(creatorProfile.user._id !== userId){
+        const error = new Error('Unauthorized')
+        error.status = 401
+        return next(error)
+      }
+
       res.status = 200
       res.json(creatorProfile)
     })
     .catch((error) => next(error))
-
   })
   .delete('/creator/:id', authenticate, (req, res, next) => {
     const id = req.params.id
-    console.log(id)
-    CreatorProfile.findByIdAndRemove(id)
-    .then((creatorProfile) => {
-      if (!creatorProfile) { return next() }
-      res.status = 204
-    })
+    userId = req.account._id
+
+    CreatorProfile.findById(id)
+      .then((creatorProfile) => {
+        if (!creatorProfile) { return next() }
+
+        if(creatorProfile.user._id !== userId){
+          const error = new Error('Unauthorized')
+          error.status = 401
+          return next(error)
+        }
+
+        CreatorProfile.findByIdAndRemove(id)
+        .then((creatorProfile) => {
+          if (!creatorProfile) { return next() }
+          res.status = 204
+        })
+        .catch((error) => next(error))
+      })
     .catch((error) => next(error))
   })
